+++ src/file-map-url-loader.ts
@@ -12,9 +12,7 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {UrlLoader} from 'polymer-analyzer';
-import {parseUrl} from 'polymer-analyzer/lib/core/utils';
-import {ResolvedUrl} from 'polymer-analyzer/lib/model/url';
+import {ResolvedUrl, UrlLoader} from 'polymer-analyzer';
 
 import {getFileContents} from './streams';
 
@@ -25,10 +23,10 @@ import File = require('vinyl');
  * that have been gathered by a `BuildBundler` transform stream.
  */
 export class FileMapUrlLoader implements UrlLoader {
-  files: Map<string, File>;
+  files: Map<ResolvedUrl, File>;
   fallbackLoader?: UrlLoader;
 
-  constructor(files: Map<string, File>, fallbackLoader?: UrlLoader) {
+  constructor(files: Map<ResolvedUrl, File>, fallbackLoader?: UrlLoader) {
     this.files = files;
     this.fallbackLoader = fallbackLoader;
   }
@@ -43,7 +41,7 @@ export class FileMapUrlLoader implements UrlLoader {
   // Try to load the file from the map.  If not in the map, try to load
   // from the fallback loader.
   async load(url: ResolvedUrl): Promise<string> {
-    const file = this.files.get(parseUrl(url).pathname || '/')!;
+    const file = this.files.get(url);
 
     if (file == null) {
       if (this.fallbackLoader) {