+++ src/test/bundle_test.ts
@@ -15,7 +15,8 @@
 /// <reference path="../../node_modules/@types/mocha/index.d.ts" />
 
 import {assert} from 'chai';
-import {Bundle} from 'polymer-bundler/lib/bundle-manifest';
+import {FsUrlResolver, PackageRelativeUrl, ResolvedUrl} from 'polymer-analyzer';
+import {Bundle} from 'polymer-bundler';
 import {ProjectConfig, ProjectOptions} from 'polymer-project-config';
 
 import File = require('vinyl');
@@ -27,6 +28,7 @@ const mergeStream = require('merge-stream');
 
 import {BuildAnalyzer} from '../analyzer';
 import {BuildBundler, Options as BuildBundlerOptions} from '../bundle';
+import {LocalFsPath} from '../path-transformers';
 import {AsyncTransformStream} from '../streams';
 
 const defaultRoot = path.resolve('test-fixtures/bundler-data');
@@ -43,6 +45,11 @@ class FileTransform extends AsyncTransformStream<File, {}> {
       this.transform(this, file.clone());
     }
   }
+}
+
+function resolveUrls(root: string, urls: string[]): ResolvedUrl[] {
+  const resolver = new FsUrlResolver(root);
+  return urls.map((u) => resolver.resolve(u as PackageRelativeUrl)!);
 }
 
 suite('BuildBundler', () => {
@@ -69,7 +76,7 @@ suite('BuildBundler', () => {
     }
     bundledStream = bundledStream.pipe(bundler);
     bundler = new BuildBundler(config, analyzer);
-    files = new Map();
+    files = new Map<LocalFsPath, File>();
     bundledStream.on('data', (file: File) => {
       files.set(file.path, file);
     });
@@ -368,7 +375,7 @@ suite('BuildBundler', () => {
   test('bundler outputs html imports that are not inlined', async () => {
     await setupTest(
         {root: defaultRoot, entrypoint: 'entrypoint-only.html'},
-        {excludes: ['framework.html']});
+        {excludes: resolveUrls(defaultRoot, ['framework.html'])});
     // We should have an entrypoint-only.html file (bundled).
     assert.isOk(getFile('entrypoint-only.html'));
     // We should have the html import that was excluded from inlining.
@@ -407,12 +414,15 @@ suite('BuildBundler', () => {
       root: 'test-fixtures/test-project',
       entrypoint: 'index.html',
       fragments: ['shell.html'],
+      componentDir: 'bower_components',
     };
 
     test('excludes: html file urls listed are not inlined', async () => {
-      await setupTest(
-          projectOptions,
-          {excludes: ['bower_components/loads-external-dependencies.html']});
+      await setupTest(projectOptions, {
+        excludes: resolveUrls(
+            projectOptions.root,
+            ['bower_components/loads-external-dependencies.html']),
+      });
       assert.isOk(
           getFile('bower_components/loads-external-dependencies.html'),
           'Excluded import is passed through the bundler');
@@ -422,7 +432,9 @@ suite('BuildBundler', () => {
     });
 
     test('excludes: html files in folders listed are not inlined', async () => {
-      await setupTest(projectOptions, {excludes: ['bower_components/']});
+      await setupTest(
+          projectOptions,
+          {excludes: resolveUrls(projectOptions.root, ['bower_components/'])});
       assert.include(
           getFileOrDie('shell.html'),
           '<link rel="import" href="bower_components/dep.html">');
@@ -492,7 +504,7 @@ suite('BuildBundler', () => {
         // Custom strategy creates a separate bundle for everything in the
         // `bower_components` folder.
         strategy: (bundles) => {
-          const bowerBundle = new Bundle();
+          const bowerBundle = new Bundle('html-fragment');
           bundles.forEach((bundle) => {
             bundle.files.forEach((file) => {
               if (file.includes('bower_components')) {
@@ -511,11 +523,18 @@ suite('BuildBundler', () => {
     });
 
     test('urlMapper: fn(), applies bundle url mapper function', async () => {
+      const urlResolver = new FsUrlResolver(projectOptions.root);
       await setupTest(projectOptions, {
-        urlMapper: (bundles) => {
-          const map = new Map<string, Bundle>();
+        urlMapper: (bundles: Bundle[]) => {
+          const map = new Map<ResolvedUrl, Bundle>();
           for (const bundle of bundles) {
-            map.set(`bundled/${Array.from(bundle.entrypoints)}`, bundle);
+            map.set(
+                urlResolver.resolve(
+                    `bundled/${
+                            [...bundle.entrypoints]
+                                .map((u) => urlResolver.relative(u))
+                                .join()}` as PackageRelativeUrl)!,
+                bundle);
           }
           return map;
         }